Move and document matrix multiplication program into "Matrix Operations"
- Move Program to Multiply Two Matrices by Passing Matrix to a Function.c into Matrix Operations folder.
- Add inline documentation explaining each function, parameters, input validation, and multiplication logic.
- Remove duplicate top-level file.
- Add related LeetCode references in file comments for learning links.
