--- Program to Multiply Two Matrices by Passing Matrix to a Function.c
@@ -1,74 +1,0 @@
-#include <stdio.h>
-
-#define MAX 10  // Maximum size of matrix
-
-// Function to multiply two matrices
-void multiplyMatrices(int first[MAX][MAX], int second[MAX][MAX], int result[MAX][MAX], int r1, int c1, int c2) {
-    // Initializing result matrix elements to 0
-    for (int i = 0; i < r1; ++i) {
-        for (int j = 0; j < c2; ++j) {
-            result[i][j] = 0;
-        }
-    }
-
-    // Multiplying firstMatrix and secondMatrix and storing in result
-    for (int i = 0; i < r1; ++i) {
-        for (int j = 0; j < c2; ++j) {
-            for (int k = 0; k < c1; ++k) {
-                result[i][j] += first[i][k] * second[k][j];
-            }
-        }
-    }
-}
-
-// Function to display a matrix
-void displayMatrix(int matrix[MAX][MAX], int row, int col) {
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
-            printf("%d ", matrix[i][j]);
-        }
-        printf("\n");
-    }
-}
-
-int main() {
-    int first[MAX][MAX], second[MAX][MAX], result[MAX][MAX];
-    int r1, c1, r2, c2;
-
-    printf("Enter rows and columns for first matrix: ");
-    scanf("%d %d", &r1, &c1);
-
-    printf("Enter rows and columns for second matrix: ");
-    scanf("%d %d", &r2, &c2);
-
-    // Check if multiplication is possible
-    if (c1 != r2) {
-        printf("Error! Column of first matrix is not equal to row of second matrix.\n");
-        return 0;
-    }
-
-    // Input first matrix
-    printf("Enter elements of first matrix:\n");
-    for (int i = 0; i < r1; ++i) {
-        for (int j = 0; j < c1; ++j) {
-            scanf("%d", &first[i][j]);
-        }
-    }
-
-    // Input second matrix
-    printf("Enter elements of second matrix:\n");
-    for (int i = 0; i < r2; ++i) {
-        for (int j = 0; j < c2; ++j) {
-            scanf("%d", &second[i][j]);
-        }
-    }
-
-    // Multiply the two matrices
-    multiplyMatrices(first, second, result, r1, c1, c2);
-
-    // Display the result
-    printf("\nResultant Matrix:\n");
-    displayMatrix(result, r1, c2);
-
-    return 0;
-}
